backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic-settings==2.3.4
python-jose==3.3.0

# Auth / hashing (FIX)
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
httpx==0.27.0
pdfplumber==0.11.0
python-docx==1.1.2
openpyxl==3.1.5
pillow==10.3.0
pytest==8.2.2
psycopg[binary]


requests==2.32.3
PyJWT==2.9.0
